updating visualizations plugin
