+++ src/legacy/core_plugins/visualizations/public/index.ts
@@ -26,22 +26,7 @@
 // @ts-ignore Used only by tsvb, vega, input control vis
 export { defaultFeedbackMessage } from 'ui/vis/default_feedback_message';
 // @ts-ignore
-export { visFactory } from 'ui/np_ready/public/legacy/vis_factory';
-// @ts-ignore
 export { DefaultEditorSize } from 'ui/vis/editor_size';
-
-/**
- * Legacy types which haven't been moved to this plugin yet, but
- * should be eventually.
- *
- * @public
- */
-import * as types from 'ui/np_ready/public/legacy/vis';
-export type Vis = types.Vis;
-export type VisParams = types.VisParams;
-export type VisState = types.VisState;
-export { VisualizationController } from 'ui/np_ready/public/types/vis_type';
-export { Status } from 'ui/np_ready/public/legacy/update_status';
 
 /**
  * Static np-ready code, re-exported here so consumers can import from
