+++ src/legacy/core_plugins/visualizations/public/np_ready/public/filters/index.ts
@@ -17,4 +17,4 @@
  * under the License.
  */
 
-export * from './filters_service';
+export * from './vis_filters';
