+++ src/legacy/core_plugins/visualizations/public/np_ready/public/types/index.ts
@@ -18,3 +18,7 @@
  */
 
 export * from './types_service';
+// @ts-ignore
+export { BaseVisType } from './base_vis_type';
+// @ts-ignore
+export { ReactVisType } from './react_vis_type';
